+++ packages/vest/src/hooks/optional/omitOptionalFields.ts
@@ -9,7 +9,6 @@ import {
 import { Events } from 'BusEvents';
 import type { IsolateSuite } from 'IsolateSuite';
 import { IsolateTest } from 'IsolateTest';
-import { useEmit } from 'Runtime';
 import { TestWalker } from 'TestWalker';
 
 /**
@@ -20,7 +19,7 @@ import { TestWalker } from 'TestWalker';
 
 export function useOmitOptionalFields(): void {
   const root = VestRuntime.useAvailableRoot<IsolateSuite>();
-  const emit = useEmit();
+  const emit = VestRuntime.useEmit();
 
   const optionalFields = root?.getOptionalFields();
