+++ packages/vest/src/suite/createSuite.ts
@@ -3,7 +3,7 @@ import { assign, CB } from 'vest-utils';
 
 import { Events } from 'BusEvents';
 import { IsolateSuite } from 'IsolateSuite';
-import { useCreateVestState, useEmit, usePrepareEmitter } from 'Runtime';
+import { useCreateVestState } from 'Runtime';
 import { SuiteContext } from 'SuiteContext';
 import {
   SuiteName,
@@ -12,6 +12,7 @@ import {
   TGroupName,
 } from 'SuiteResultTypes';
 import { Suite } from 'SuiteTypes';
+import { useInitVestBus } from 'VestBus';
 import { getTypedMethods } from 'getTypedMethods';
 import { useCreateSuiteResult } from 'suiteResult';
 import { useSuiteRunResult } from 'suiteRunResult';
@@ -47,7 +48,7 @@ function createSuite<
   function suite(...args: Parameters<T>): SuiteRunResult<F, G> {
     return SuiteContext.run({}, () => {
       // eslint-disable-next-line vest-internal/use-use
-      const emit = useEmit();
+      const emit = VestRuntime.useEmit();
 
       emit(Events.SUITE_RUN_STARTED);
 
@@ -61,15 +62,17 @@ function createSuite<
   // We do this within the VestRuntime so that the suite methods
   // will be bound to the suite's stateRef and be able to access it.
   return VestRuntime.Run(stateRef, () => {
+    useInitVestBus();
+
     return assign(
       // We're also binding the suite to the stateRef, so that the suite
       // can access the stateRef when it's called.
       VestRuntime.persist(suite),
       {
         get: VestRuntime.persist(useCreateSuiteResult),
-        remove: usePrepareEmitter<string>(Events.REMOVE_FIELD),
-        reset: usePrepareEmitter(Events.RESET_SUITE),
-        resetField: usePrepareEmitter<string>(Events.RESET_FIELD),
+        remove: VestRuntime.usePrepareEmitter<string>(Events.REMOVE_FIELD),
+        reset: VestRuntime.usePrepareEmitter(Events.RESET_SUITE),
+        resetField: VestRuntime.usePrepareEmitter<string>(Events.RESET_FIELD),
         ...bindSuiteSelectors<F, G>(VestRuntime.persist(useCreateSuiteResult)),
         ...getTypedMethods<F, G>(),
       }
@@ -82,7 +85,7 @@ function useRunSuiteCallback<
   F extends TFieldName,
   G extends TGroupName
 >(suiteCallback: T, ...args: Parameters<T>): () => SuiteRunResult<F, G> {
-  const emit = useEmit();
+  const emit = VestRuntime.useEmit();
 
   return () => {
     suiteCallback(...args);
